+++ src/main/components/AnotherComponent.java
@@ -1,12 +1,11 @@
 package main.components;
 
 
-import main.annotations.Component;
-import main.annotations.PostConstructor;
+import main.container.annotations.Component;
+import main.container.annotations.PostConstructor;
 
 @Component
-public
-class AnotherComponent {
+public class AnotherComponent {
     @PostConstructor
     public void init() {
         System.out.println("AnotherComponent initialized");
