+++ src/main/components/NotAComponent.java
@@ -1,0 +1,4 @@
+package main.components;
+
+public class NotAComponent {
+}
