+++ src/main/container/DependencyInjectionContainer.java
@@ -1,9 +1,11 @@
-package main;
+package main.container;
 
-import main.annotations.Autowired;
-import main.annotations.Component;
-import main.annotations.PostConstructor;
-import main.annotations.Qualifier;
+import main.container.services.ClassScanner;
+import main.container.annotations.Autowired;
+import main.container.annotations.Component;
+import main.container.annotations.PostConstructor;
+import main.container.annotations.Qualifier;
+import main.container.services.ComponentPreProcessor;
 
 import java.lang.annotation.*;
 import java.lang.reflect.*;
@@ -13,6 +15,8 @@ import java.util.concurrent.*;
 
 public class DependencyInjectionContainer {
     private final Map<Class<?>, Object> instances = new ConcurrentHashMap<>();
+    private final List<ComponentPreProcessor> preProcessors = new ArrayList<>();
+
 
     public void scanAndRegisterComponents(String basePackage) throws Exception {
         ClassScanner classScanner = new ClassScanner();
@@ -27,11 +31,14 @@ public class DependencyInjectionContainer {
     }
 
     public <T> void register(Class<T> componentClass) throws Exception {
+        //Will skip not components for now instead of exception;
         if (!componentClass.isAnnotationPresent(Component.class)) {
-            throw new IllegalArgumentException("Class must be annotated with @Component: " + componentClass);
+            return;
+//            throw new IllegalArgumentException("Class must be annotated with @Component: " + componentClass);
         }
 
         T instance = createInstance(componentClass);
+        preProcess(instance);
         postProcess(instance);
 
         instances.put(componentClass, instance);
@@ -112,6 +119,17 @@ public class DependencyInjectionContainer {
         }
     }
 
+    //Возможность добавлять обработчики ПЕРЕД добавлением в контейнер
+    private <T> void preProcess(T component) {
+        for (ComponentPreProcessor preProcessor : preProcessors) {
+            preProcessor.process(component);
+        }
+    }
+
+    //Добавить препроцессор к кмопоненту
+    public void addComponentPreProcessor(ComponentPreProcessor preProcessor) {
+        preProcessors.add(preProcessor);
+    }
 
     //Возможность добавлять обработчики ПОСЛЕ добавления в контейнер
     private void postProcess(Object instance) {
