+++ src/main/container/services/ComponentPreProcessor.java
@@ -1,0 +1,6 @@
+package main.container.services;
+
+//Возможность добавлять обработчики ПЕРЕД добавлением в контейнер
+public interface ComponentPreProcessor {
+    void process(Object component);
+}
