+++ src/main/container/services/CustomPreProcessor.java
@@ -1,0 +1,8 @@
+package main.container.services;
+
+public class CustomPreProcessor implements ComponentPreProcessor {
+    @Override
+    public void process(Object component) {
+        System.out.println("Pre process for " + component.getClass());
+    }
+}
